+++ src/actions/monitoring.actions.ts
@@ -7,10 +7,19 @@ import { CreditModule } from '@/types/models/current-control/credit-module';
 export async function getMonitoring() {
   const response = await campusFetch<Sheet>('monitoring');
 
+  if (!response.ok) {
+    throw new Error(`${response.status} Error`);
+  }
+
   return response.json();
 }
 
 export async function getMonitoringById(id: string) {
   const response = await campusFetch<CreditModule>(`monitoring/${id}`);
+
+  if (!response.ok) {
+    throw new Error(`${response.status} Error`);
+  }
+
   return response.json();
 }
