+++ src/app/[locale]/(private)/module/studysheet/[id]/components/event-plan-table.tsx
@@ -3,14 +3,15 @@ import React from 'react';
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/components/ui/table';
 import { useTranslations } from 'next-intl';
 import { EventsPlan } from '@/types/models/current-control/events-plan';
-import { LecturerItemCell } from '@/app/[locale]/(private)/module/studysheet/[id]/components/LecturerItemCell';
+import { LecturerItemCell } from '@/app/[locale]/(private)/module/studysheet/[id]/components/lecturer-item-cell';
 
 interface Props {
   eventsPlan: EventsPlan[];
-  t: ReturnType<typeof useTranslations>;
 }
 
-export function EventPlanTable({ eventsPlan, t }: Props) {
+export function EventPlanTable({ eventsPlan }: Props) {
+  const t = useTranslations('private.study-sheet.table');
+
   return (
     <Table>
       <TableHeader>
