+++ src/app/[locale]/(private)/module/studysheet/[id]/components/lecturer-item-cell.tsx
