+++ src/app/[locale]/(private)/module/studysheet/[id]/constants.ts
@@ -1,4 +1,4 @@
-export enum SHEET_TRANSLATION_KEYS {
+export enum SheetTranslationKeys {
   JOURNAL = 'journal',
   EVENT_PLAN = 'event-plan',
   INTERNAL_MATERIALS = 'internal-materials',
