--- src/app/[locale]/(private)/module/studysheet/constants.ts
@@ -1,1 +1,0 @@
-export const INTL_NAMESPACE = 'private.study-sheet';
