+++ src/app/[locale]/(private)/module/studysheet/page.tsx
@@ -1,15 +1,14 @@
 import { getTranslations } from 'next-intl/server';
-import { Studysheet } from '@/app/[locale]/(private)/module/studysheet/components/Studysheet';
-import { INTL_NAMESPACE } from '@/app/[locale]/(private)/module/studysheet/constants';
+import { StudySheet } from '@/app/[locale]/(private)/module/studysheet/components/study-sheet';
 
 export async function generateMetadata({ params: { locale } }: { params: { locale: string } }) {
-  const t = await getTranslations({ locale, namespace: INTL_NAMESPACE });
+  const t = await getTranslations({ locale, namespace: 'private.study-sheet' });
 
   return {
     title: t('title'),
   };
 }
 
-export default async function StudysheetPage() {
-  return <Studysheet />;
+export default async function StudySheetPage() {
+  return <StudySheet />;
 }
