+++ src/types/enums/current-control/semester.ts
@@ -1,4 +1,4 @@
-export enum SEMESTER {
+export enum Semester {
   ALL = 'all',
   FIRST = '1',
   SECOND = '2',
